requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.1
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-braces==1.13.0
django-cors-middleware==1.3.1
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jwt==0.5.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.3.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.9.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24
whitenoise==4.1.2
wrapt==1.10.11
